Changing `StaticCodeAnalysisService` to a general `StaticCodeAnalyzer`
